+++ docs/search/all_d.js
@@ -3,6 +3,6 @@ var searchData=
   ['handle_0',['Handle',['../class_crab_u_i_1_1_c_u_i_slider.html#a6cf1635561b9e4c32a7c542552e06a9b',1,'CrabUI::CUISlider']]],
   ['hidechildrenoutsideframe_1',['HideChildrenOutsideFrame',['../class_crab_u_i_1_1_c_u_i_component.html#a809e2ca8b575f4532ebf47285fd4760e',1,'CrabUI::CUIComponent']]],
   ['how_20layout_20is_20calculated_2',['How layout is calculated',['../_howlayoutiscalculated.html',1,'index']]],
-  ['how_20to_20use_20them_3',['How to use them',['../_data_flow.html#autotoc_md18',1,'']]],
+  ['how_20to_20use_20them_3',['How to use them',['../_data_flow.html#autotoc_md20',1,'']]],
   ['hydrate_4',['Hydrate',['../class_crab_u_i_1_1_c_u_i_component.html#a5ea833cc01691fd22bd719bd9a2f4aac',1,'CrabUI::CUIComponent']]]
 ];