+++ docs/search/all_f.js
@@ -2,11 +2,11 @@ var searchData=
 [
   ['layout_20is_20calculated_0',['How layout is calculated',['../_howlayoutiscalculated.html',1,'index']]],
   ['layout_20props_1',['Layout Props',['../_layout_props.html',1,'index']]],
-  ['layouts_2',['Layouts',['../_overview.html#autotoc_md23',1,'\subpage Layouts'],['../_layouts.html',1,'Layouts']]],
+  ['layouts_2',['Layouts',['../_overview.html#autotoc_md25',1,'\subpage Layouts'],['../_layouts.html',1,'Layouts']]],
   ['layouttest_3',['LayoutTest',['../class_crab_u_i_test_1_1_layout_test.html',1,'CrabUITest']]],
-  ['link_20crabui_20cuimaincomponent_20main_20component_20endlink_4',['@link CrabUI.CUIMainComponent Main Component @endlink',['../_overview.html#autotoc_md24',1,'']]],
+  ['link_20crabui_20cuimaincomponent_20main_20component_20endlink_4',['@link CrabUI.CUIMainComponent Main Component @endlink',['../_overview.html#autotoc_md26',1,'']]],
   ['load_5',['Load',['../class_crab_u_i_1_1_c_u_i_main_component.html#a8364f7e94fb25d5717f8125ce2df1d75',1,'CrabUI::CUIMainComponent']]],
   ['log_6',['Log',['../class_crab_u_i_1_1_c_u_i.html#aa0ce4751fde5325adf2675c44a320b73',1,'CrabUI::CUI']]],
   ['lua_7',['Usage in Lua',['../_usagein_lua.html',1,'']]],
-  ['lua_8',['Same example in lua',['../_usagein_lua.html#autotoc_md31',1,'']]]
+  ['lua_8',['Same example in lua',['../_usagein_lua.html#autotoc_md33',1,'']]]
 ];