+++ docs/search/pages_0.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['0_0',['0.2.1.0',['../md__c_sharp_2_client_2_crab_u_i_2_changelog.html',1,'']]],
-  ['0_202_201_200_1',['0.2.1.0',['../md__c_sharp_2_client_2_crab_u_i_2_changelog.html',1,'']]]
+  ['0_0',['0.2.2.0',['../md__c_sharp_2_client_2_crab_u_i_2_changelog.html',1,'']]],
+  ['0_202_202_200_1',['0.2.2.0',['../md__c_sharp_2_client_2_crab_u_i_2_changelog.html',1,'']]]
 ];