+++ docs/search/pages_1.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['1_200_0',['0.2.1.0',['../md__c_sharp_2_client_2_crab_u_i_2_changelog.html',1,'']]]
+  ['2_200_0',['0.2.2.0',['../md__c_sharp_2_client_2_crab_u_i_2_changelog.html',1,'']]],
+  ['2_202_200_1',['0.2.2.0',['../md__c_sharp_2_client_2_crab_u_i_2_changelog.html',1,'']]]
 ];