+++ docs/search/pages_5.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['data_20flow_0',['Commands / Data flow',['../_data_flow.html',1,'index']]]
+  ['flow_0',['Commands / Data flow',['../_data_flow.html',1,'index']]]
 ];