+++ docs/search/pages_6.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['flow_0',['Commands / Data flow',['../_data_flow.html',1,'index']]]
+  ['graphic_20props_0',['Graphic Props',['../_graphic_props.html',1,'index']]]
 ];