+++ docs/search/pages_7.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['graphic_20props_0',['Graphic Props',['../_graphic_props.html',1,'index']]]
+  ['how_20layout_20is_20calculated_0',['How layout is calculated',['../_howlayoutiscalculated.html',1,'index']]]
 ];