+++ docs/search/pages_d.js
@@ -1,5 +1,4 @@
 var searchData=
 [
-  ['serialization_0',['Serialization',['../_serialization.html',1,'index']]],
-  ['styles_20palettes_1',['Styles, Palettes',['../_stylespalettes.html',1,'index']]]
+  ['tree_0',['Creating a tree',['../_creatingatree.html',1,'index']]]
 ];